Data type defined in lib type, sample data moved to sampleData file
